# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Project 2 API",
    "version": "1.0.0",
    "description": "OAuth auth, validation, CRUD for items and orders."
  },
  "servers": [
    { "url": "http://localhost:8080" },
    { "url": "https://yourapp.onrender.com" }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "openid": "OpenID scope",
              "email": "Access to email",
              "profile": "Access to profile"
            }
          }
        }
      }
    },
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "sku": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "currency": { "type": "string" },
          "inStock": { "type": "boolean" },
          "tags": { "type": "array", "items": { "type": "string" } }
        },
        "required": ["title", "sku", "price"]
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderNumber": { "type": "string" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "itemId": { "type": "string" },
                "quantity": { "type": "integer" }
              },
              "required": ["itemId", "quantity"]
            }
          },
          "total": { "type": "number" },
          "currency": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "paid", "shipped", "cancelled"] },
          "customerEmail": { "type": "string" }
        },
        "required": ["orderNumber", "items", "total", "customerEmail"]
      }
    }
  },
  "paths": {
    "/api/items": {
      "get": {
        "summary": "List items",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create item (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/items/{id}": {
      "get": {
        "summary": "Get item",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Update item (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Validation error" }, "401": { "description": "Unauthorized" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete item (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" }, "401": { "description": "Unauthorized" }, "404": { "description": "Not found" } }
      }
    },
    "/api/orders": {
      "get": {
        "summary": "List orders",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create order (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "Get order",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Update order (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Order" } } } },
        "responses": { "200": { "description": "Updated" }, "400": { "description": "Validation error" }, "401": { "description": "Unauthorized" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete order (protected)",
        "security": [{ "oauth2": ["openid"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" }, "401": { "description": "Unauthorized" }, "404": { "description": "Not found" } }
      }
    }
  }
}